+++ src/main/java/study/step3/models/Lottos.java
@@ -1,4 +1,4 @@
-package study.step3;
+package study.step3.models;
 
 import java.util.List;
