+++ src/main/java/study/step3/models/Winners.java
@@ -1,4 +1,6 @@
-package study.step3;
+package study.step3.models;
+
+import study.step3.Rank;
 
 import java.util.List;
