+++ src/main/java/study/step3/views/InputView.java
@@ -1,4 +1,4 @@
-package study.step3;
+package study.step3.views;
 
 import java.util.Scanner;
