+++ src/test/java/study/step3/models/LottosTest.java
@@ -1,17 +1,17 @@
-package study.step3;
+package study.step3.models;
 
 import org.junit.jupiter.api.Test;
+import study.step3.Rank;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static study.step3.constants.CommonConstants.PRICE_PER_LOTTO;
 
-public class WinnersTest {
+public class LottosTest {
     @Test
-    void 수익률_계산하기() {
+    void 당첨_확인하기() {
         Numbers winLottoNumbers = new Numbers(Arrays.asList(4, 5, 6, 7, 8, 9));
         List<Lotto> lottoList = new ArrayList<>(
                 Arrays.asList(
@@ -22,10 +22,9 @@ public class WinnersTest {
         List<Lotto> rankedLottoList = lottos.rankLottos(winLottoNumbers);
 
         Winners winners = new Winners(rankedLottoList);
-        Money inputMoney = new Money(String.valueOf(lottoList.size() * PRICE_PER_LOTTO));
-        double earningRate = winners.earningRate(inputMoney);
 
-        // 총 상금 예상 금액 1555000
-        assertThat(earningRate).isEqualTo(inputMoney.divide(1555000));
+        assertThat(winners.numberOfRankers(Rank.FOURTH)).isEqualTo(1);
+        assertThat(winners.numberOfRankers(Rank.THIRD)).isEqualTo(1);
+        assertThat(winners.numberOfRankers(Rank.SECOND)).isEqualTo(1);
     }
 }
