+++ src/test/java/study/step3/models/MoneyTest.java
@@ -1,4 +1,4 @@
-package study.step3;
+package study.step3.models;
 
 import org.junit.jupiter.api.Test;
 import study.step3.exception.LottoInputMoneyTypeException;
